requirements.txt
Flask==2.3.3
Pillow==10.1.0
numpy==1.24.3
torch==2.1.2
torchvision==0.16.2
gunicorn==21.2.0
Werkzeug==2.3.7